Added a half of the AOE skills with the implementation pending on the combat actually starting. Just continuing to add more backend functionality

+++ app/src/main/java/chimera/epic_rpg_thing/view/newCharacterActivity.java
@@ -750,7 +750,7 @@ public class newCharacterActivity extends AppCompatActivity {
         return finalHealth;
     }
 
-    public void setFinalHealth(short finalHealth) {
+    public void setFinalHealth(int finalHealth) {
         this.finalHealth = finalHealth;
     }
